Refactor redux store
