+++ src/index.js
@@ -1,25 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import { Provider } from "react-redux";
+import { BrowserRouter } from "react-router-dom";
+
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
-import { Provider } from "react-redux";
-import { createStore, applyMiddleware, compose } from "redux";
-import thunk from "redux-thunk";
-import { BrowserRouter } from "react-router-dom";
-
-import rootReducer from "./reducers";
-import addTokenToLocalStorage from "./middleware/add-token-to-local-storage";
-
-const store = createStore(
-  rootReducer,
-  compose(
-    applyMiddleware(thunk, addTokenToLocalStorage),
-    window.__REDUX_DEVTOOLS_EXTENSION__
-      ? window.__REDUX_DEVTOOLS_EXTENSION__()
-      : f => f
-  )
-);
+import store from "./store";
 
 ReactDOM.render(
   <Provider store={store}>
