+++ src/middleware/add-token-to-local-storage.js
@@ -2,11 +2,11 @@ import { LOGIN_SUCCESS } from "../actions";
 
 const addTokenToLocalStorage = _store => next => action => {
   if (action.type === LOGIN_SUCCESS) {
-    const token = action.payload;
+    const token = action.payload.token;
     localStorage.setItem("token", token);
   }
 
-  next(action);
+  next({ type: action.type });
 };
 
 export default addTokenToLocalStorage;
