+++ src/middleware/index.js
@@ -1,0 +1,4 @@
+import addUserToLocalStorage from "./add-user-to-local-storage";
+import addTokenToLocalStorage from "./add-token-to-local-storage";
+
+export { addUserToLocalStorage, addTokenToLocalStorage };
